+++ src/axom/primal/geometry/Polygon.hpp
@@ -13,6 +13,7 @@
 #define AXOM_PRIMAL_POLYGON_HPP_
 
 #include "axom/core/Array.hpp"
+#include "axom/core/StaticArray.hpp"
 #include "axom/primal/geometry/Point.hpp"
 #include "axom/primal/geometry/Vector.hpp"
 
@@ -25,39 +26,6 @@ namespace primal
 namespace
 {
 static constexpr int DEFAULT_MAX_NUM_VERTICES = 8;
-
-/*!
- * \brief A helper class for the Polygon class
- *
- * This class helps define a StackArray with a current size and additional
- * functions to mirror axom::Array.
- */
-template <typename T, int N>
-struct StaticArray : public StackArray<T, N>
-{
-  AXOM_HOST_DEVICE int size() const { return m_size; }
-  AXOM_HOST_DEVICE void push_back(const T& obj)
-  {
-    assert(m_size < N);
-    if(m_size < N)
-    {
-      StackArray<T, N>::m_data[m_size++] = obj;
-    }
-  }
-
-  AXOM_HOST_DEVICE void clear()
-  {
-    for(T& datum : StackArray<T, N>::m_data)
-    {
-      datum = T();
-    }
-    m_size = 0;
-  }
-
-private:
-  int m_size {0};
-};
-
 } /* end anonymous namespace */
 
 /**
@@ -106,7 +74,7 @@ public:
   // axom::Array for dynamic array type, StaticArray for static array type
   using ArrayType = std::conditional_t<ARRAY_TYPE == PolygonArray::Dynamic,
                                        axom::Array<PointType>,
-                                       StaticArray<PointType, MAX_VERTS>>;
+                                       axom::StaticArray<PointType, MAX_VERTS>>;
 
 public:
   /// Default constructor for an empty polygon (dynamic array specialization).
@@ -121,6 +89,45 @@ public:
   template <PolygonArray P_ARRAY_TYPE = ARRAY_TYPE,
             std::enable_if_t<P_ARRAY_TYPE == PolygonArray::Static, int> = 0>
   AXOM_HOST_DEVICE Polygon()
+  { }
+
+  /*!
+   * \brief Destructor for Polygon. Suppress CUDA warnings for
+   *        dynamic axom::Array.
+   */
+  AXOM_SUPPRESS_HD_WARN
+  AXOM_HOST_DEVICE
+  ~Polygon() { m_vertices.clear(); }
+
+  /*!
+   * \brief Copy assignment operator for Polygon. Suppress CUDA warnings for
+   *        dynamic axom::Array.
+   */
+  AXOM_SUPPRESS_HD_WARN
+  AXOM_HOST_DEVICE
+  Polygon& operator=(const Polygon& other)
+  {
+    if(this == &other)
+    {
+      return *this;
+    }
+
+    m_vertices = other.m_vertices;
+    return *this;
+  }
+
+  /// Copy constructor for Polygon. Specializations are necessary to
+  /// remove __host__ __device__ warning for axom::Array usage with
+  /// the dynamic array type.
+  template <PolygonArray P_ARRAY_TYPE = ARRAY_TYPE,
+            std::enable_if_t<P_ARRAY_TYPE == PolygonArray::Dynamic, int> = 0>
+  Polygon(const Polygon& other) : m_vertices(other.m_vertices)
+  { }
+
+  /// Copy constructor for Polygon (static array specialization)
+  template <PolygonArray P_ARRAY_TYPE = ARRAY_TYPE,
+            std::enable_if_t<P_ARRAY_TYPE == PolygonArray::Static, int> = 0>
+  AXOM_HOST_DEVICE Polygon(const Polygon& other) : m_vertices(other.m_vertices)
   { }
 
   /*!
@@ -175,7 +182,17 @@ public:
   AXOM_HOST_DEVICE
   int numVertices() const { return static_cast<int>(m_vertices.size()); }
 
-  /// Appends a vertex to the list of vertices
+  /*!
+   * \brief Appends a vertex to the list of vertices
+   *
+   * \param [in] pt the point to be appended to the list of vertices
+   *
+   * \note If the array type is static and the list of vertices is full,
+   *       addVertex will not modify the list of vertices.
+   *
+   * \sa axom::StaticArray::push_back() for behavior when array type is static
+   *     and the list of vertices is full.
+   */
   AXOM_HOST_DEVICE
   void addVertex(const PointType& pt) { m_vertices.push_back(pt); }
 
@@ -229,6 +246,22 @@ public:
     }
 
     return normal;
+  }
+
+  /// \brief Reverses orientation of the polygon in-place
+  AXOM_HOST_DEVICE
+  void reverseOrientation()
+  {
+    const int nverts = numVertices();
+    const int mid = nverts >> 1;
+
+    // Swap leftmost/rightmost vertices, midpoint unchanged
+    for(int i = 0; i < mid; ++i)
+    {
+      const int left = i;
+      const int right = nverts - i - 1;
+      axom::utilities::swap(m_vertices[left], m_vertices[right]);
+    }
   }
 
   /*!